move endpoint requests into peer utilities
